Add IValidatator and move X12ValidationHierarchy
